One thread, one poll: 80% CPU reduction
Profiling shows a mutex inside libndi, so polling in two threads doesn't
help us as far as libndi is concerned (but theoretically might for the
OBS side).

This effectively meant that every ms, two threads were fighting for a
mutex.

Now that we have everything in one thread, we can also reduce the time
we poll for, as it's a max time (timeout).

This takes me from 2 cores at 100% + a third at 50% to one core at 50%