base: Add aux output helper
Uses obs_view_t to pipe a source to an output. Currently only being
used for preview output but this will make it very easy to create a
full "aux output manager" which can create additional outputs and
send a source's video through it.

Audio is TBD. There's not really tools right now like obs_view_t in
OBS for doing the same thing with audio. Might make the configuration
just allow the user to select an audio track to use.