logging/config: log levels, blog->obs_log, & command-line arg tweaks
I am changing from direct `blog(...)` calls to the template "recommended" `obs_log(...)`.
This cleans up the code a little bit to get rid of the redundant "[DistroAV] " prefixes.
This also lets the code gate the log levels a little better.
I have waffled on doing this several times but I am finally biting the bullet on this.

blog already uses va_list, so this is not much worse.
The only "worse" part is the dynamically allocated and freed buffer used by snprintf.
Hopefully this does not thrash memory too much.