Implement "Keep Last Received Frame" functionality
This is a new checkbox that is enabled by default. When it is enabled it will keep the last frame that was received on the source, no matter if the sender NDI instance was disconnected or is sending "none" frame data.
When it is disabled, it will wait for up to 3 seconds to receive a new frame and, if there is no new frame, will clear it on the receiver end.

The way this is possible is via. a new helper method called `is_filter_invalid` (probably not the final name) which is checking the source width and height, as well as if the filter is enabled on the source and if the source is active anywhere.
Sources don't update when they aren't rendered in the main OBS view from what I can tell (e.g. a Browser Source would just seem like it's frozen) so this is in-line with the current way OBS handles it