Fix filter chain rendering
We now check if both the target (filter above our own, OR parent if we're in first place) and the actual filter parent are valid instead of just the target.

Additionally, we get the source width and height for the filter, which if it's in first place will give the original source width and height (otherwise it would've picked the parent's and that will be with all filters *afterwards* applied, such as Crop or Aspect Ratio)

Also, I moved the known width and height check / update before the texrender logic.

Inside the render call we check if the target is the parent and if so, we render the filter via. `obs_source_skip_video_filter`, otherwise we just use `obs_source_video_render` on the target.