Convert offscreen render method to sync
Offscreen isn't really necessary - sync also works for async sources.

This fixes an issue with e.g. Color Correction (and potentially other filters) being above the Dedicated NDI Output filter and rendering everything black.

Additionally, the `video_output_lock_frame` method was put in an inner `gs_stagesurface_map` call, as it returns a boolean on whether or not it was able to map it.

Performance seems to be on-par with the offscreen method, though with the stagesurface_map change I was getting slightly better performance (at least when I last tested it a few weeks ago for the other PR)